ArchRepo 1.0 beta1 release
Squashed commit of the following:

commit 698cbcac3d3f1caeabe417095f4fd67276cc9e30
Author: fantix <fantix@exoweb.net>
Date:   Sat Sep 8 11:16:48 2012 +0800

    Using zeromq for ipc

commit 42e5e247afb9966a6501d0e3542c8b425f7d2338
Author: fantix <fantix@exoweb.net>
Date:   Thu Sep 6 18:00:32 2012 +0800

    Prepare for using zeromq

commit 39eb01467a26ddf7a99a017159a9bdcf4676fb77
Author: Fantix King <fantix.king@gmail.com>
Date:   Mon Sep 3 23:29:38 2012 +0800

    INSTALL.txt and LICENSE.txt, added archrepo_sync.py (TODO: lock)

commit d83bc4b1db5b0a07c851ba1c20d2c876de3f2e93
Author: Fantix King <fantix.king@gmail.com>
Date:   Sun Sep 2 22:55:36 2012 +0800

    bug fixes, ready to work on package detail page

+++ bin/archrepo_serve.py
@@ -10,12 +10,19 @@ if __name__ == '__main__':
                    help='Output verbose, 0 (default) for silent, 1 for info, '
                         '2 for everything')
     args = p.parse_args()
-    if args.verbose == 1:
+    if args.verbose == 0:
+        logging.basicConfig(level=logging.ERROR)
+        logging.error('Logging is set to ERROR')
+    elif args.verbose == 1:
         logging.basicConfig(level=logging.INFO)
         logging.info('Logging is set to INFO')
     elif args.verbose == 2:
         logging.basicConfig(level=logging.DEBUG)
         logging.debug('Logging is set to DEBUG')
+    import cherrypy
+    cherrypy.log.screen = False
+    cherrypy.log.access_log.level = cherrypy.log.access_log.parent.level
+    cherrypy.log.error_log.level = cherrypy.log.error_log.parent.level
 
     from archrepo.db_pool import buildPool
     from archrepo.repo import Processor
@@ -24,10 +31,13 @@ if __name__ == '__main__':
     pool = buildPool()
     p = Processor(pool=pool)
     p.serve()
-    web_server = ArchRepoWebServer(pool)
-    try:
-        web_server.serve_forever()
-    except KeyboardInterrupt:
-        pass
-    finally:
-        p.kill()
+    if p.serving:
+        web_server = ArchRepoWebServer(pool)
+        try:
+            web_server.serve_forever()
+        except KeyboardInterrupt:
+            pass
+        finally:
+            p.kill()
+    else:
+        logging.critical('Another ArchRepo processor is working, try again later')
